first test introduced for new it
